requirements.txt
altgraph==0.17
future==0.18.2
importlib-metadata==4.8.1
infi.systray==0.1.12
pefile==2021.9.3
pyinstaller==4.5.1
pyinstaller-hooks-contrib==2021.3
pywin32-ctypes==0.2.0
typing-extensions==3.10.0.2
zipp==3.5.0
